requirements-dev.txt
pytest
flake8
python-semantic-release
pre-commit
commitizen
